requirements.txt
# LLaMA Fine-Tuning Dependencies
# Install with: pip install -r requirements.txt

# Core ML frameworks
torch>=2.1.0
transformers>=4.36.0
datasets>=2.16.0
accelerate>=0.25.0

# Parameter-efficient fine-tuning
peft>=0.7.0
bitsandbytes>=0.41.0

# Training utilities
trl>=0.7.0

# Unsloth for faster training (optional but recommended)
# Install separately: pip install "unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git"

# Data processing
pyyaml>=6.0
tqdm>=4.66.0

# Logging and monitoring
wandb>=0.16.0

# For Colab
huggingface_hub>=0.20.0
